requirements.txt
alembic==1.7.7
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.0
Flask==2.1.0
Flask-Bcrypt==0.7.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
password==0.2
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.32
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
zipp==3.7.0

requests~=2.27.1